- Return tap UI and core methods if TapTalk is not initialized

+++ taptalk/src/main/java/io/taptalk/TapTalk/Helper/TapTalk.java
@@ -312,14 +312,18 @@ public class TapTalk {
     }
 
     public static void initializeGooglePlacesApiKey(String apiKey) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         Places.initialize(appContext, apiKey);
     }
 
-    private static void checkTapTalkInitialized() {
+    public static boolean checkTapTalkInitialized() {
         if (null == tapTalk) {
-            throw new IllegalStateException(appContext.getString(R.string.tap_init_taptalk));
+            Log.e(getClientAppName(), appContext.getString(R.string.tap_init_taptalk));
+            return false;
         }
+        return true;
     }
 
     public static void setLoggingEnabled(boolean enabled) {
@@ -345,7 +349,9 @@ public class TapTalk {
      */
 
     public static void authenticateWithAuthTicket(String authTicket, boolean connectOnSuccess, TapCommonListener listener) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         if (null == authTicket || "".equals(authTicket)) {
             listener.onError(ERROR_CODE_INVALID_AUTH_TICKET, ERROR_MESSAGE_INVALID_AUTH_TICKET);
         } else {
@@ -419,12 +425,16 @@ public class TapTalk {
     }
 
     public static boolean isAuthenticated() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return false;
+        }
         return TAPDataManager.getInstance().checkAccessTokenAvailable();
     }
 
     public static void logoutAndClearAllTapTalkData(TapCommonListener listener) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         TAPDataManager.getInstance().logout(new TAPDefaultDataView<TAPCommonResponse>() {
             @Override
             public void onSuccess(TAPCommonResponse response) {
@@ -445,7 +455,9 @@ public class TapTalk {
     }
 
     public static void clearAllTapTalkData() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         TAPDataManager.getInstance().deleteAllPreference();
         TAPDataManager.getInstance().deleteAllFromDatabase();
         TAPDataManager.getInstance().deleteAllManagerData();
@@ -462,8 +474,10 @@ public class TapTalk {
      */
 
     public static void connect(TapCommonListener listener) {
-        checkTapTalkInitialized();
-        if (isAuthenticated()) {
+        if (!checkTapTalkInitialized()) {
+            return;
+
+        }        if (isAuthenticated()) {
             TAPConnectionManager.getInstance().connect(listener);
         } else {
             listener.onError(ERROR_CODE_ACCESS_TOKEN_UNAVAILABLE, ERROR_MESSAGE_ACCESS_TOKEN_UNAVAILABLE);
@@ -471,22 +485,30 @@ public class TapTalk {
     }
 
     public static void disconnect() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         TAPConnectionManager.getInstance().close(NOT_CONNECTED);
     }
 
     public static boolean isConnected() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return false;
+        }
         return TAPConnectionManager.getInstance().getConnectionStatus() == CONNECTED;
     }
 
     public static void setAutoConnectEnabled(boolean enabled) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         isAutoConnectDisabled = !enabled;
     }
 
     public static boolean isAutoConnectEnabled() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return false;
+        }
         return !isAutoConnectDisabled;
     }
 
@@ -497,38 +519,49 @@ public class TapTalk {
      */
 
     public static int getClientAppIcon() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return -1;
+        }
         return clientAppIcon;
     }
 
     public static String getClientAppName() {
-        checkTapTalkInitialized();
         return clientAppName;
     }
 
     public static TapTalkImplementationType getTapTalkImplementationType() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return null;
+        }
         return implementationType;
     }
 
     public static void updateApplicationBadgeCount() {
-        checkTapTalkInitialized();
-        if (isAuthenticated()) TAPNotificationManager.getInstance().updateUnreadCount();
+        if (!checkTapTalkInitialized()) {
+            return;
+
+        }        if (isAuthenticated()) TAPNotificationManager.getInstance().updateUnreadCount();
     }
 
     // TODO: 22 August 2019 CORE MODEL
     public static Map<String, String> getCoreConfigs() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return null;
+        }
         return new HashMap<>(coreConfigs);
     }
 
     public static Map<String, String> getProjectConfigs() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return null;
+        }
         return new HashMap<>(projectConfigs);
     }
 
     public static Map<String, String> getCustomConfigs() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return null;
+        }
         return new HashMap<>(customConfigs);
     }
 
@@ -576,12 +609,16 @@ public class TapTalk {
     }
 
     public static void setAutoContactSyncEnabled(boolean enabled) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         isAutoContactSyncDisabled = !enabled;
     }
 
     public static boolean isAutoContactSyncEnabled() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return false;
+        }
         return !isAutoContactSyncDisabled;
     }
 
@@ -592,12 +629,16 @@ public class TapTalk {
      */
 
     public static TAPUserModel getTaptalkActiveUser() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return null;
+        }
         return TAPChatManager.getInstance().getActiveUser();
     }
 
     public static void refreshActiveUser(TapCommonListener listener) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         new Thread(() -> {
             if (null != TAPChatManager.getInstance().getActiveUser()) {
                 TAPDataManager.getInstance().getUserByIdFromApi(TAPChatManager.getInstance().getActiveUser().getUserID(), new TAPDefaultDataView<TAPGetUserResponse>() {
@@ -664,28 +705,38 @@ public class TapTalk {
      */
 
     public static List<TapListener> getTapTalkListeners() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return null;
+        }
         return tapTalk.tapListeners;
     }
 
     private static void addTapTalkListener(TapListener listener) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         tapTalk.tapListeners.add(listener);
     }
 
     private static void setTapTalkScreenOrientation(TapTalkScreenOrientation orientation) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         TapTalk.screenOrientation = orientation;
     }
 
     private static TapTalkScreenOrientation getTapTalkScreenOrientation() {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return null;
+        }
         return TapTalk.screenOrientation;
     }
 
     // Enable/disable in-app notification after chat fragment goes inactive or to background
     private static void setInAppNotificationEnabled(boolean enabled) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         TAPNotificationManager.getInstance().setRoomListAppear(!enabled);
     }
 
@@ -694,7 +745,9 @@ public class TapTalk {
 //    }
 
     private void createAndShowBackgroundNotification(Context context, int notificationIcon, Class destinationClass, TAPMessageModel newMessageModel) {
-        checkTapTalkInitialized();
+        if (!checkTapTalkInitialized()) {
+            return;
+        }
         TAPNotificationManager.getInstance().createAndShowBackgroundNotification(context, notificationIcon, destinationClass, newMessageModel);
     }
